refactor(workflow): split parity extractor into honest halves
the phase 1 exit criterion said delete engine-contract.ts and re-point the
parity suite at the manifests. analysis of all 1,118 lines showed the premise
was ~85% wrong: ~950 lines scrape ENGINE processor source (setNodeVariable
writes, node.data reads, outputHandle emissions) — information a NodeManifest
does not carry, from processors a web test process cannot execute (bullmq/
redis/db imports). that scraping's real retirement is the planned
resolvability suite, which executes processors against fixtures instead.

so: engine-contract.ts is deleted and split into
- engine-write-scrape.ts — the engine-side scraper, functionally unchanged,
  with a header documenting why manifests don't obsolete it and naming the
  resolvability suite as its retirement path
- builder-declared-keys.ts — the declared-shape reader, now catalog-first for
  migrated types (that half genuinely moved to the catalog); stays textual
  because ts interfaces aren't runtime-reflectable and zod schemas are
  deliberately narrower than the data interfaces
- monorepo-paths.ts — shared path/source helpers

the one true manifest win: enum body text-parsing replaced by importing
WorkflowNodeType directly. zero assertions changed — the reverse
written-subset-of-advertised drift catcher runs identically; no allowlist
entries added or removed.
